move functions from common data file to timelogs data file and update path imports
